requirements.txt
Qt5.py
PySide2
